Use less memory for ContextImpl.mServiceInitializationStateArray
`ContextImpl.{mServiceInitializationStateArray, mServiceCache}` takes
up memory that scales with # of services, # of contexts, and # of
apps in the system.

`ContextImpl.mServiceCache` has to be kept for some app compat issues
(some apps using reflection to access this field - see b/75964116)

However, `ContextImpl.mServiceInitializationStateArray` can use
byte instead of int, which would save 3/4 of the memory it takes
up today.

Bug: 457769621
Flag: EXEMPT pure_refactor
Test: atest CtsContentTestCases:android.content.cts.ContextTest

+++ core/java/android/app/ContextImpl.java
@@ -385,10 +385,10 @@ class ContextImpl extends Context {
     @UnsupportedAppUsage
     final Object[] mServiceCache = SystemServiceRegistry.createServiceCache();
 
-    static final int STATE_UNINITIALIZED = 0;
-    static final int STATE_INITIALIZING = 1;
-    static final int STATE_READY = 2;
-    static final int STATE_NOT_FOUND = 3;
+    static final byte STATE_UNINITIALIZED = 0;
+    static final byte STATE_INITIALIZING = 1;
+    static final byte STATE_READY = 2;
+    static final byte STATE_NOT_FOUND = 3;
 
     /** @hide */
     @IntDef(prefix = { "STATE_" }, value = {
@@ -405,7 +405,7 @@ class ContextImpl extends Context {
      * {@link #STATE_INITIALIZING} or {@link #STATE_READY},
      */
     @ServiceInitializationState
-    final int[] mServiceInitializationStateArray = new int[mServiceCache.length];
+    final byte[] mServiceInitializationStateArray = new byte[mServiceCache.length];
 
     private final Object mDeviceIdListenerLock = new Object();
     /**
